update:Alter data field entity submission-service & feedback-service

+++ services/submission-service/src/main/java/com/submission_service/model/dto/request/SubmissionRequest.java
@@ -10,8 +10,9 @@ import org.springframework.web.multipart.MultipartFile;
 @Setter
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class SubmissionRequest {
-    Integer studentId;
-    Integer assignmentId;
+    Long studentId;
+    Long assignmentId;
     String studentPlantUmlCode;
+
 //    private MultipartFile plantUmlFile;
 }
